+++ crates/simulate/src/agent/mod.rs
@@ -5,11 +5,15 @@
 //! An abstract representation of an agent on the EVM, to be used in simulations.
 //! Some examples of agents are market makers or arbitrageurs.
 //! All agents must implement the [`Agent`] trait.
-use std::{cell::{RefCell, RefMut}, rc::Rc};
 use bytes::Bytes;
-use revm::primitives::{Address, ExecutionResult, Log, TransactTo, TxEnv, B160, U256};
+use ethers::abi::Tokenize;
+use revm::primitives::{Address, ExecutionResult, Log, Output, TransactTo, TxEnv, B160, U256};
+use std::{
+    cell::{RefCell, RefMut},
+    rc::Rc,
+};
 
-use crate::environment::{IsDeployed, SimulationContract, SimulationEnvironment};
+use crate::environment::{IsDeployed, NotDeployed, SimulationContract, SimulationEnvironment};
 
 pub mod admin;
 /// Describes the gas settings for a transaction.
@@ -58,14 +62,63 @@ pub trait Agent {
         }
     }
 
-    // TODO: Not sure `read_logs` needs to be mutable self.
     /// Gets the most current event (which is all that is stored in the event buffer).
-    fn read_logs(&mut self) -> Vec<Log> {
+    fn read_logs(&self) -> Vec<Log> {
         self.simulation_environment()
             .event_buffer
             .read()
             .unwrap()
             .to_vec()
+    }
+    /// Deploy a contract to the current simulation environment.
+    fn deploy<T: Tokenize>(
+        &self,
+        contract: SimulationContract<NotDeployed>,
+        args: T,
+    ) -> SimulationContract<IsDeployed> {
+        // Append constructor args (if available) to generate the deploy bytecode.
+        let constructor = contract.base_contract.abi().constructor();
+
+        let bytecode = match constructor {
+            Some(constructor) => Bytes::from(
+                constructor
+                    .encode_input(contract.bytecode.clone(), &args.into_tokens())
+                    .unwrap(),
+            ),
+            None => Bytes::from(contract.bytecode.clone()),
+        };
+
+        // Take the execution result and extract the contract address.
+        // Manager address will always be the sender for contract deployments.
+
+        let deploy_transaction = self.build_deploy_transaction(bytecode);
+        let execution_result = self.simulation_environment().execute(deploy_transaction);
+        let output = match execution_result {
+            ExecutionResult::Success { output, .. } => output,
+            ExecutionResult::Revert { output, .. } => panic!("Failed due to revert: {:?}", output),
+            ExecutionResult::Halt { reason, .. } => panic!("Failed due to halt: {:?}", reason),
+        };
+        let contract_address = match output {
+            Output::Create(_, address) => address.unwrap(),
+            _ => panic!("failed"),
+        };
+
+        contract.to_deployed(contract_address)
+    }
+
+    fn build_deploy_transaction(&self, bytecode: Bytes) -> TxEnv {
+        TxEnv {
+            caller: self.address(),
+            gas_limit: self.transact_settings().gas_limit,
+            gas_price: self.transact_settings().gas_price,
+            gas_priority_fee: None,
+            transact_to: TransactTo::create(),
+            value: U256::ZERO,
+            data: bytecode,
+            chain_id: None,
+            nonce: None,
+            access_list: Vec::new(),
+        }
     }
 }
