+++ crates/simulate/src/manager.rs
@@ -2,7 +2,7 @@
 //! Simulation managers are used to manage the environments for a simulation.
 //! Managers are responsible for adding agents, running agents, deploying contracts, calling contracts, and reading logs.
 
-use std::{collections::HashMap, cell::RefCell, rc::Rc};
+use std::{cell::RefCell, collections::HashMap, rc::Rc};
 
 use bytes::Bytes;
 use revm::primitives::{ExecutionResult, Output};
